Refactor DataView into per-concern hooks and components
DataView.tsx had grown to ~1075 lines spanning six concerns. Split it into
a DataView/ directory (mirroring the ProtocolCarousel pattern):

- useDataViewUrlState: filters/sort state + URL query-string sync
- useSessionQuery: query params, pagination, filtersKey, stale-guarded fetch
- useSessionSelection: page/all-matching selection model + filter reset
- useSessionMutations: bulk export and delete flows
- useDataViewColumns: column definitions + SortHeader
- DataViewToolbar: status chips, search, filter popover, bulk actions
- DataView.tsx: ~250-line orchestrator composing the hooks

Behavior-preserving; the dataViewUrlState codec and its tests moved into
the directory unchanged. No context provider: all state composes at one
call site, so hooks keep the data flow explicit.
